test(snapshot): split test_snapshot_service.py to resolve LOC blocker
Split test_snapshot_service.py (616 lines, custom limit 500) into three
logical test files to comply with CI LOC limits:

- test_snapshot_lookup.py (189 lines): find_snapshot_by_branch tests (8 tests)
- test_snapshot_baselines.py (183 lines): baseline operation tests (5 tests)
- test_snapshot_service.py (205 lines): core ops + metadata tests (7 tests)

Shared snapshot_dir fixture moved to conftest.py (121 lines) for reuse.

All files now under 300-line warning threshold and 400-line CI block threshold.
All 20 tests pass with same coverage.
